Format everything

+++ src/ZuniswapV2Library.sol
@@ -64,7 +64,11 @@ library ZuniswapV2Library {
         );
     }
 
-    function getAmountOut(uint256 amountIn, uint256 reserveIn, uint256 reserveOut) public pure returns (uint256) {
+    function getAmountOut(
+        uint256 amountIn,
+        uint256 reserveIn,
+        uint256 reserveOut
+    ) public pure returns (uint256) {
         if (amountIn == 0) revert InsufficientAmount();
         if (reserveIn == 0 || reserveOut == 0) revert InsufficientLiquidity();
 
@@ -75,14 +79,22 @@ library ZuniswapV2Library {
         return numerator / denominator;
     }
 
-    function getAmountsOut(address factory, uint256 amountIn, address[] memory path) public returns (uint256[] memory) {
+    function getAmountsOut(
+        address factory,
+        uint256 amountIn,
+        address[] memory path
+    ) public returns (uint256[] memory) {
         if (path.length < 2) revert InvalidPath();
         uint256[] memory amounts = new uint256[](path.length);
         amounts[0] = amountIn;
 
-        for(uint i; i < path.length-1; i++) {
-            (uint256 reserve0, uint256 reserve1) = getReserves(factory, path[i], path[i+1]);
-            amounts[i+1] = getAmountOut(amounts[i], reserve0, reserve1);
+        for (uint256 i; i < path.length - 1; i++) {
+            (uint256 reserve0, uint256 reserve1) = getReserves(
+                factory,
+                path[i],
+                path[i + 1]
+            );
+            amounts[i + 1] = getAmountOut(amounts[i], reserve0, reserve1);
         }
 
         return amounts;
